+++ src/Papercut.Core/AppConstants.cs
@@ -2,28 +2,34 @@
 // 
 // Copyright © 2008 - 2012 Ken Robertson
 // Copyright © 2013 - 2020 Jaben Cargman
-//  
+// 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//  
+// 
 // http://www.apache.org/licenses/LICENSE-2.0
-//  
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License. 
+// limitations under the License.
 
-// ReSharper disable once CheckNamespace
-// want this to be in this namespace to match KeyValuePair namespace
-namespace System.Collections.Generic
+
+namespace Papercut.Core
 {
-    public static class KeyValuePair
+    using System;
+    using System.IO;
+
+    public static class AppConstants
     {
-        public static KeyValuePair<TKey, TValue> Create<TKey, TValue>(TKey key, TValue value)
-        {
-            return new KeyValuePair<TKey, TValue>(key, value);
-        }
+        public const string ApplicationName = "Papercut SMTP";
+
+        public const string CompanyName = "Changemaker Studios";
+
+        public static string DataDirectory { get; } =
+            Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData),
+                CompanyName,
+                ApplicationName);
     }
 }