+++ src/Papercut.Infrastructure.IPComm/Network/PapercutIPCommProtocol.cs
@@ -20,9 +20,11 @@ namespace Papercut.Infrastructure.IPComm.Network
     using System.IO;
     using System.Net.Sockets;
 
+    using Common;
     using Common.Domain;
     using Common.Extensions;
 
+    using Core;
     using Core.Infrastructure.Json;
     using Core.Infrastructure.MessageBus;
 
@@ -47,7 +49,7 @@ namespace Papercut.Infrastructure.IPComm.Network
             this.Connection = connection;
             this.Logger.ForContext("ConnectionId", this.Connection.Id);
 
-            this.Connection.SendLine("PAPERCUT").Wait();
+            this.Connection.SendLine(AppConstants.ApplicationName.ToUpper()).Wait();
         }
 
         protected override void ProcessRequest(string incomingRequest)