+++ src/Papercut.UI/ViewModels/ForwardViewModel.cs
@@ -22,6 +22,10 @@ namespace Papercut.ViewModels
 
     using Caliburn.Micro;
 
+    using Common;
+
+    using Core;
+
     using Papercut.Properties;
 
     public class ForwardViewModel : Screen
@@ -118,7 +122,7 @@ namespace Papercut.ViewModels
             {
                 MessageBox.Show(
                     "All the text boxes are required, fill them in please.",
-                    "Papercut",
+                    AppConstants.ApplicationName,
                     MessageBoxButton.OK,
                     MessageBoxImage.Warning);
                 return;
@@ -128,7 +132,7 @@ namespace Papercut.ViewModels
             {
                 MessageBox.Show(
                     "You need to enter valid email addresses.",
-                    "Papercut",
+                    AppConstants.ApplicationName,
                     MessageBoxButton.OK,
                     MessageBoxImage.Warning);
                 return;