+++ src/Papercut.UI/ViewModels/MainViewModel.cs
@@ -31,6 +31,10 @@ namespace Papercut.ViewModels
 
     using Caliburn.Micro;
 
+    using Common;
+
+    using Core;
+
     using ICSharpCode.AvalonEdit.Utils;
 
     using MahApps.Metro.Controls;
@@ -55,7 +59,7 @@ namespace Papercut.ViewModels
         IHandle<ShowMainWindowEvent>,
         IHandle<ShowOptionWindowEvent>
     {
-        const string WindowTitleDefault = "Papercut";
+        const string WindowTitleDefault = AppConstants.ApplicationName;
 
         readonly ForwardRuleDispatch _forwardRuleDispatch;
 
@@ -110,10 +114,7 @@ namespace Papercut.ViewModels
 
         public string LogText
         {
-            get
-            {
-                return _logText;
-            }
+            get => _logText;
             set
             {
                 _logText = value;
@@ -123,10 +124,7 @@ namespace Papercut.ViewModels
 
         public bool IsDeactivated
         {
-            get
-            {
-                return _isDeactivated;
-            }
+            get => _isDeactivated;
             set
             {
                 _isDeactivated = value;
@@ -136,10 +134,7 @@ namespace Papercut.ViewModels
 
         public string WindowTitle
         {
-            get
-            {
-                return _windowTitle;
-            }
+            get => _windowTitle;
             set
             {
                 _windowTitle = value;
@@ -154,7 +149,7 @@ namespace Papercut.ViewModels
             return productVersion.Split('+').FirstOrDefault();
         }
 
-        public string Version => $"Papercut v{GetVersion()}";
+        public string Version => $"{AppConstants.ApplicationName} v{GetVersion()}";
 
         public bool IsLogOpen
         {
@@ -179,6 +174,36 @@ namespace Papercut.ViewModels
                     _isDeleteAllConfirmOpen = value;
                     NotifyOfPropertyChange(() => IsDeleteAllConfirmOpen);
                 }
+            }
+        }
+
+        public int MainWindowHeight
+        {
+            get => _mainWindowHeight;
+            set
+            {
+                if (value == _mainWindowHeight) return;
+
+                // ignore non-normal window sizes
+                if (_window.WindowState != WindowState.Normal) return;
+
+                _mainWindowHeight = value;
+                NotifyOfPropertyChange(() => MainWindowHeight);
+            }
+        }
+
+        public int MainWindowWidth
+        {
+            get => _mainWindowWidth;
+            set
+            {
+                if (value == _mainWindowWidth) return;
+
+                // ignore non-normal window sizes
+                if (_window.WindowState != WindowState.Normal) return;
+
+                _mainWindowWidth = value;
+                NotifyOfPropertyChange(() => MainWindowWidth);
             }
         }
 
@@ -259,6 +284,8 @@ namespace Papercut.ViewModels
 
         public Deque<string> _currentLogHistory = new Deque<string>();
         private bool _isDeleteAllConfirmOpen;
+        private int _mainWindowWidth;
+        private int _mainWindowHeight;
 
         void SetupObservables()
         {